Moved custom icons and locations to the locaitonmanager package
